CompactCardView and MiniCardView refactor #612
