+++ WMIAdventure/frontend/src/js/components/battle/atoms/UserStatistic/UserStatistic.js
