+++ WMIAdventure/frontend/src/js/components/battle/atoms/UserStatistic/index.js
