+++ WMIAdventure/frontend/src/js/components/battle/atoms/UserStatistic/styled-components/ValueDiv.js
